[SWM-269] Feat : shadow effect on gnb
